move layout.svelte functions to utils
